feat: add sports sidebar
Instead of fetching the `sports` as a nested entry, and then just making
some stupid grouping, we spread the relationship of the `sports` table.
Since the `sports` table is static, and has properties that do not
overlap with the `user_sports`, the spread makes more sense and
simplifies the data, removing such need for grouping and so.